1. 片段编辑体验优化 2. 部分拉伸组件逻辑优化 3. zookeeper协议，新增认证重复问题修复 4. 其他细节处理

+++ src/main/java/cn/oyzh/easyshell/fx/zk/ShellZKHistoryDataTableView.java
@@ -88,8 +88,10 @@ public class ShellZKHistoryDataTableView extends FXTableView<ShellZKHistoryData>
      * @param data 数据
      */
     public void deleteData(ShellZKHistoryData data) {
-        ShellZKDataUtil.deleteHistory(this.nodePath, data.getSaveTime(), this.client);
-        this.removeItem(data);
+        if (MessageBox.confirm(I18nHelper.deleteHistory())) {
+            ShellZKDataUtil.deleteHistory(this.nodePath, data.getSaveTime(), this.client);
+            this.removeItem(data);
+        }
     }
 
     /**
